+++ apps/web/src/app/[locale]/(app)/(space)/(dashboard)/events/page.tsx
@@ -1,7 +1,7 @@
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import type { Metadata } from "next";
 import { getTranslation } from "@/i18n/server";
-import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
+import { createPrivateSSRHelper } from "@/trpc/next/ssr";
 import { EventsPage } from "./events-page";
 import { eventsSearchParamsSchema } from "./schema";
