+++ apps/web/src/app/[locale]/(app)/(space)/settings/billing/page.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { getTranslation } from "@/i18n/server";
 import { isFeatureEnabled } from "@/lib/feature-flags/server";
-import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
+import { createPrivateSSRHelper } from "@/trpc/next/ssr";
 import { BillingPageClient } from "./page-client";
 
 export default async function BillingSettingsPage() {
