+++ apps/web/src/app/[locale]/(app)/(space)/settings/calendars/page.tsx
@@ -19,7 +19,7 @@ import {
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
 import { isFeatureEnabled } from "@/lib/feature-flags/server";
-import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
+import { createPrivateSSRHelper } from "@/trpc/next/ssr";
 import { CalendarConnectionList } from "./components/calendar-connection-list";
 import { ConnectCalendarDropdown } from "./components/connect-calendar-dropdown";
 import { DefaultCalendarSelect } from "./components/default-calendar-select";
