+++ apps/web/src/app/[locale]/(app)/(space)/settings/members/page.tsx
@@ -1,7 +1,7 @@
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import type { Metadata } from "next";
 import { getTranslation } from "@/i18n/server";
-import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
+import { createPrivateSSRHelper } from "@/trpc/next/ssr";
 import { MembersSettingsPageClient } from "./page-client";
 
 export default async function MembersSettingsPage() {
