+++ apps/web/src/app/[locale]/(app)/(space)/settings/spaces/page.tsx
@@ -11,7 +11,7 @@ import {
 } from "@/app/components/settings-layout";
 import { Trans } from "@/i18n/client";
 import { getTranslation } from "@/i18n/server";
-import { createPrivateSSRHelper } from "@/trpc/server/create-ssr-helper";
+import { createPrivateSSRHelper } from "@/trpc/next/ssr";
 import { SpacesList } from "./components/spaces-list";
 
 export default async function Page() {
